# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop API Documentation",
        "description": "This is the API documentation for Petshop.",
        "contact": {
            "name": "Godwin (DeGod)",
            "email": "support@petshop-api.test"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/product/{uuid}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Fetch a product",
                "operationId": "746fa4f8de5977e09f27ece56cf6541e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Edit an existing product",
                "operationId": "acdd4c26a753de845e262faaedee56a9",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "category_uuid",
                                    "title",
                                    "price",
                                    "description",
                                    "metadata"
                                ],
                                "properties": {
                                    "category_uuid": {
                                        "description": "Category UUID",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "title": {
                                        "description": "Product title",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "price": {
                                        "description": "Product price",
                                        "type": "number",
                                        "example": ""
                                    },
                                    "description": {
                                        "description": "Product description",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "metadata": {
                                        "description": "Product metadata",
                                        "properties": {
                                            "brand": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete an existing product",
                "operationId": "81bbb2d16e8242115d638b14c2fd6d70",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}